Replace sanitize(val) with quote(val, col_definition)
Non standard columns like Point, Hstore are not handled correctly using
ActiveRecord::Base.sanitize. For example sanitize of a Point gives
=> "'---\n- 77.086565\n- 28.455375\n'"

Instead use ActiveRecord::Base.connection.quote with the actual column
definition to correctly serialize the value for insertion. In the above
example usage of quote with the definition for the column will give:
=> "'(77.086565,28.455375)'"